Begin transferring exceptions to separate file
